requirements.txt
Django
Pillow
psycopg2-binary
djangorestframework
python-decouple
djangorestframework-simplejwt
django-countries
drf-yasg
celery
redis

django-cors-headers
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
googleapis-common-protos
python-twitter
oauthlib
httplib2
facebook-sdk